requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.86.0,<2.0.0
langchain==0.3.26
langchain-openai==0.3.28
langchain-community==0.3.26
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langsmith>=0.3.45,<1.0.0 
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.26.4
pandas==2.0.3
streamlit==1.28.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
